+++ app/components/Popover.tsx
@@ -62,7 +62,7 @@ const Popover: React.FC<Props> = ({
   }
 
   return (
-    <ReakitPopover {...rest} hideOnEsc={false} hideOnClickOutside>
+    <StyledPopover {...rest} hideOnEsc={false} hideOnClickOutside>
       <Contents
         $shrink={shrink}
         $width={width}
@@ -71,7 +71,7 @@ const Popover: React.FC<Props> = ({
       >
         {children}
       </Contents>
-    </ReakitPopover>
+    </StyledPopover>
   );
 };
 
@@ -82,6 +82,10 @@ type ContentsProps = {
   $scrollable: boolean;
   $mobilePosition?: "top" | "bottom";
 };
+
+const StyledPopover = styled(ReakitPopover)`
+  z-index: ${depths.modal};
+`;
 
 const Contents = styled.div<ContentsProps>`
   display: ${(props) => (props.$flex ? "flex" : "block")};
