+++ shared/i18n/locales/en_US/translation.json
@@ -214,7 +214,6 @@
   "{{ count }} member": "{{ count }} member",
   "{{ count }} member_plural": "{{ count }} members",
   "Group members": "Group members",
-  "Icon": "Icon",
   "Show menu": "Show menu",
   "Choose icon": "Choose icon",
   "Loading": "Loading",
@@ -488,10 +487,11 @@
   "You can edit the name and other details at any time, however doing so often might confuse your team mates.": "You can edit the name and other details at any time, however doing so often might confuse your team mates.",
   "Name": "Name",
   "Sort": "Sort",
-  "Collections are for grouping your documents. They work best when organized around a topic or internal team — Product or Engineering for example.": "Collections are for grouping your documents. They work best when organized around a topic or internal team — Product or Engineering for example.",
-  "This is the default level of access, you can give individual users or groups more access once the collection is created.": "This is the default level of access, you can give individual users or groups more access once the collection is created.",
+  "Collections are used to group documents and choose permissions": "Collections are used to group documents and choose permissions",
+  "Permission": "Permission",
+  "The default access for workspace members, you can share with more users or groups later.": "The default access for workspace members, you can share with more users or groups later.",
   "Public document sharing": "Public document sharing",
-  "When enabled any documents within this collection can be shared publicly on the internet.": "When enabled any documents within this collection can be shared publicly on the internet.",
+  "Allow documents within this collection to be shared publicly on the internet.": "Allow documents within this collection to be shared publicly on the internet.",
   "Create": "Create",
   "{{ groupName }} was added to the collection": "{{ groupName }} was added to the collection",
   "Could not add user": "Could not add user",
